break out unsynced sync service tests
